--- packages/components/src/Calendar/index.js
@@ -1,5 +1,0 @@
-export { default as Calendar } from './Calendar';
-export { default as CalendarNav } from './components/CalendarNav';
-export { default as CalendarMonth } from './components/CalendarMonth';
-export * from './components/CalendarDays';
-export * from './components/CalendarWeekdays';
