+++ packages/components/src/DatePicker/index.js
@@ -1,0 +1,1 @@
+export { default } from './DatePicker';
